requirements.txt
streamlit>=1.47.0
firebase-admin>=7.1.0
pyrebase4>=4.8.0
streamlit-authenticator>=0.4.2
pandas>=2.0.0
numpy>=1.24.0
plotly>=6.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
reportlab>=4.0.0